Rename rules/syntax_error to rules/error
Fixes clippy warning for modules with same names as their
parents.
