+++ src/rules/error/syntax_error.rs
